Populate _config.yml with my info
As I write this, there have been [133 commits][1] to the Minima theme
since the latest published version. I updated to the latest commit
because this way I can easily compare what I'm seeing to the [theme
preview][2], and also because this brings a bunch of improvements:

- I can now remove my duplicated name from the bottom of the page.

  This lines up the `description` with the RSS thingy, rather than
  having some awkward extra vertical space to account for the duplicated
  `title`.

- I can now put a link to my Stack Overflow page.

I'm not really sold on the `description` in general, because it feels
weird to have the same thing go in the title of the webpage and in the
footer. I might remove it later it it feels too weird.

I'm omitting my email address because it just seems like a bad idea to
put it right there on my homepage. I mean, I have already been putting
it in all my Git commits, but I feel like that's probably more fine
because for some reason people don't seem to usually look at Git commit
messages to find email addresses.

[1]: https://github.com/jekyll/minima/compare/v2.5.1...49f6dce0727a2441f0b0c265b41b5efc7b042eb6
[2]: https://jekyll.github.io/minima/