Try to upgrade GitHub Pages Minima theme
Very sadly, GitHub Pages does not seem to want to use the [most
up-to-date version of Minima][1]. Judging from [this blog post][2], it
looks like it's possible to use any public GitHub repo theme by
replacing `theme` with `remote_theme`, so I'm going to try this to see
if it works.

This completely breaks local development as described in the README, so
I'll try to fix that if this ends up working.

[1]: f04399ddc20afa8730e839f086f0ce27aa6ce898
[2]: https://github.blog/2017-11-29-use-any-theme-with-github-pages/