Use jekyll-remote-theme to reproduce build locally
[This plugin][1] seems to be what GitHub Pages uses (since it fixes the
build), but I can't tell for sure. If it is, then this [should solve my
worries][2] about jekyll/minima@master moving around:

> You may also optionally specify a branch, tag, or commit to use by
> appending an `@` and the Git ref (e.g., `benbalter/retlab@v1.0.0` or
> `benbalter/retlab@develop`). If you don't specify a Git ref, the
> `master` branch will be used.

Fingers crossed that GitHub Pages allows this syntax too.

I'm guessing that this also makes some of the other stuff in `Gemfile`
and `Gemfile.lock` redundant, but it doesn't seem to be hurting
anything, so I'm not going to spend the effort to figure out what I can
and can't remove without breaking things.

[1]: https://github.com/benbalter/jekyll-remote-theme
[2]: https://github.com/benbalter/jekyll-remote-theme/tree/v0.4.2#declaring-your-theme