Add a picture of me to the About page
I used [Kramdown inline attributes][1] to put the photo to the side, and
used [`border-radius: 50%`][2] to make it a circle. It [looks like][3]
people usually put these sorts of things in an `assets` folder. It also
looks like [`ex` is for "relative" vertical distances][4], so I just
picked a multiple of `1ex` that made the photo look right on my laptop.
Hopefully it'll also look good on mobile.

[1]: https://talk.jekyllrb.com/t/inclusion-image-on-left-side-and-text-on-the-right-side-using-markdown-in-jekyll-site/3413
[2]: https://stackoverflow.com/a/51435610/5044950
[3]: https://stackoverflow.com/a/40198284/5044950
[4]: https://stackoverflow.com/a/2385866/5044950